Move PrefsDialog3 to gui.preftabs
